requirement.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
RPi.GPIO==0.7.1
smbus2==0.5.0
Werkzeug==3.1.3
